+++ R/lmer_pi_unstruc.R
@@ -17,8 +17,9 @@
 #' if a prediction interval or an upper or a lower prediction limit should be computed
 #' @param alpha defines the level of confidence (1-\code{alpha})
 #' @param nboot number of bootstraps
-#' @param lambda_min lower start value for bisection
-#' @param lambda_max upper start value for bisection
+#' @param delta_min lower start value for bisection
+#' @param delta_max upper start value for bisection
+#' @param tolerance tolerance for the coverage probability in the bisection
 #' @param traceplot plot for visualization of the bisection process
 #' @param n_bisec maximal number of bisection steps
 #'
@@ -60,12 +61,13 @@
 #' @importFrom graphics abline lines
 #' @importFrom lme4 fixef VarCorr bootMer
 #' @importFrom stats vcov
+#' @importFrom methods is
 #'
 #' @examples
 #' # loading lme4
 #' library(lme4)
 #'
-#' # Fitting a random effects model based on c2_dat_1
+#' # Fitting a random effects model based on c2_dat1
 #' fit <- lmer(y_ijk~(1|a)+(1|b)+(1|a:b), c2_dat1)
 #' summary(fit)
 #'
@@ -84,13 +86,14 @@ lmer_pi_unstruc <- function(model,
                     alternative="both",
                     alpha=0.05,
                     nboot=10000,
-                    lambda_min=0.01,
-                    lambda_max=10,
+                    delta_min=0.01,
+                    delta_max=10,
+                    tolerance = 1e-3,
                     traceplot=TRUE,
                     n_bisec=30){
 
         # Model must be of class lmerMod
-        if(class(model) != "lmerMod"){
+        if(!is(model, "lmerMod")){
                 stop("class(model) != lmerMod")
         }
 
@@ -371,7 +374,7 @@ lmer_pi_unstruc <- function(model,
         #----------------------------------------------------------------------
         ### Bisection
 
-        bisection <- function(f, quant_min, quant_max, n, tol = 1e-3) {
+        bisection <- function(f, quant_min, quant_max, n, tol = tolerance) {
 
 
                 c_i <- vector()
@@ -498,7 +501,7 @@ lmer_pi_unstruc <- function(model,
         }
 
         # Calculation of the degreees of freedom
-        quant_calib <- bisection(f=coverfun, quant_min=lambda_min, quant_max=lambda_max, n=n_bisec)
+        quant_calib <- bisection(f=coverfun, quant_min=delta_min, quant_max=delta_max, n=n_bisec)
 
         #----------------------------------------------------------------------