+++ NEWS.md
@@ -7,3 +7,14 @@
 * lmer_pi() is now deprecated, but three new functions are introduced for the
 computation of PI based on random effects models fit by lme4::lmer(). The new 
 functions are lmer_pi_unstruc(), lmer_pi_futvec() and lmer_pi_futmat().
+
+# predint 1.1.1
+* Argument for the tolerance in the bisection
+* Small changes in the readme
+* Bugfix in beta_bin_pi
+* Changed lambda_min and lambda_max to delta_min and delta_max in order
+        to reflect the notation of the manuscript submitted to JSS
+* Bugfix in rbbinom: rho must be bigger than 0 but smaller than 1
+* Function for the sampling of bs-data based on a random effects model fit with lme4::lmer()
+  as used in lmer_pi_futmat()
+